refactor(step1): 클래스를 vo, infrastructure 패키지로 이동시킨다.
