+++ src/test/java/kitchenpos/products/tobe/domain/vo/DisplayedName.java
@@ -1,4 +1,4 @@
-package kitchenpos.products.tobe.domain;
+package kitchenpos.products.tobe.domain.vo;
 
 import java.util.Objects;
