+++ src/test/java/kitchenpos/products/tobe/domain/vo/ProfanityName.java
@@ -1,4 +1,4 @@
-package kitchenpos.products.tobe.domain;
+package kitchenpos.products.tobe.domain.vo;
 
 import java.util.List;
 import java.util.Objects;
