+++ src/test/java/kitchenpos/products/tobe/infrastructure/JdbcTemplateProductRepository.java
@@ -1,5 +1,8 @@
-package kitchenpos.products.tobe.domain;
+package kitchenpos.products.tobe.infrastructure;
 
+import kitchenpos.products.tobe.domain.Product;
+import kitchenpos.products.tobe.domain.ProductRepository;
+import kitchenpos.products.tobe.domain.vo.ProfanityName;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
